+++ packages/parte-icons/src/Icons/ToBottomIcon.tsx
@@ -1,0 +1,37 @@
+import { convertIcon, IconProps } from "../common/Icon";
+function Icon({ size, ...rest }: Pick<IconProps, "style" | "color" | "size">) {
+  return (
+    <svg
+      width={size}
+      height={size}
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...rest}
+    >
+      <path
+        d="M8.00293 11.2998V2"
+        stroke="currentcolor"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+      <path
+        d="M12 7.33334L8 11.3333L4 7.33334"
+        stroke="currentcolor"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+      <path
+        d="M12 14H4"
+        stroke="currentcolor"
+        strokeWidth="1.5"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+}
+const ToBottomIcon = convertIcon(Icon, "to-bottom");
+export default ToBottomIcon;